Move test case to SparkPlanSuite
